requirements.txt
pandas
numpy
yfinance
plotly
streamlit
requests
beautifulsoup4
google-generativeai
statsmodels